highlighting: Move to rouge
Checked with visual diff on full page screenshot of the tmpfs and
numbers posts that nothing material changes with these updates.
`.highlight` CSS rule is removed because it makes non-{% highlight %}
indented blocks all white in the back with the new highlighter output,
and I can't remember why it was there in the first place.

The only things that seems to change is that `#include` directives have
their files greyed out since, unlike pygments, rouge doesn't output
these as <span class="cpf">, but it's not a massive deal. Looks like
it's a bug since I see that the PreprocFile is actually present in
lib/rouge/token.rb, so maybe I'll submit something to fix it later.